+++ packages/gatsby-remark-jargon/rollup.config.js
@@ -1,23 +1,25 @@
-import babel from 'rollup-plugin-babel'
-import resolve from 'rollup-plugin-node-resolve'
-import commonjs from 'rollup-plugin-commonjs'
-import json from 'rollup-plugin-json'
-import minify from 'rollup-plugin-babel-minify'
+import babel from '@rollup/plugin-babel'
+import resolve from '@rollup/plugin-node-resolve'
+import commonjs from '@rollup/plugin-commonjs'
+import json from '@rollup/plugin-json'
+import { terser } from 'rollup-plugin-terser'
 import peerDepsExternal from 'rollup-plugin-peer-deps-external'
-import { name, version, description, author, license, main, module } from './package.json'
+import { main, module, name, version, description, author, license } from './package.json'
 
 const output = [
   {
     file: main,
     format: 'cjs',
-    sourcemap: true
+    sourcemap: true,
+    exports: 'default'
   }
 ]
 if (typeof module !== 'undefined')
   output.push({
     file: module,
     format: 'es',
-    sourcemap: true
+    sourcemap: true,
+    exports: 'default'
   })
 
 export default {
@@ -25,18 +27,18 @@ export default {
   output,
   external: ['remark-jargon'],
   plugins: [
+    babel({
+      exclude: 'node_modules/**',
+      babelHelpers: 'bundled'
+    }),
     peerDepsExternal(),
     resolve(),
     commonjs(),
     json(),
-    babel({
-      exclude: 'node_modules/**',
-      plugins: ['@babel/plugin-proposal-object-rest-spread']
-    }),
-    minify({
-      comments: false,
-      sourceMap: true,
-      banner: `/**\n * ${name} | v${version}\n * ${description}\n * (c) ${new Date().getFullYear()} ${author}\n * @license ${license}\n */`
+    terser({
+      output: {
+        preamble: `/**\n * ${name} | v${version}\n * ${description}\n * (c) ${new Date().getFullYear()} ${author}\n * @license ${license}\n */`
+      }
     })
   ]
 }