+++ packages/components/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change log for: @freesewing/components
 
+## 2.8.0 (2020-08-10)
+
+### Added
+
+- Added Paco to LineDrawing component
+
+### Removed
+
+- Removed the `Footer` component
+
+## 2.7.1 (2020-07-24)
+
+### Added
+
+- Workbench now includes events debug output and enables debug by default
+
 ## 2.7.0 (2020-07-12)
 
 ### Added